pyproject.toml
[project]
name = "weather-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow>=2.7.2",
    "apache-airflow-providers-common-sql>=1.15.0",
    "apache-airflow-providers-snowflake>=5.5.1",
    "audio-recorder-streamlit>=0.0.10",
    "black>=25.1.0",
    "confluent-kafka>=2.10.1",
    "dbt-snowflake>=1.7.5",
    "isort>=6.0.1",
    "joblib>=1.5.1",
    "kafka-python>=2.0.2",
    "langchain-community>=0.3.26",
    "langchain-litellm>=0.2.1",
    "langgraph>=0.5.0",
    "pandas>=2.3.0",
    "pdf-ingest>=1.0.13",
    "pinecone>=7.3.0",
    "protobuf>=5.29.5",
    "psycopg2-binary>=2.9.10",
    "pylint>=3.3.7",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "ruff>=0.12.3",
    "scikit-learn>=1.7.0",
    "snowflake-connector-python>=3.16.0",
    "sqlfluff>=3.4.1",
    "sqlfluff-templater-dbt>=3.4.1",
    "streamlit>=1.46.1",
]
